requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
Flask
Flask-RESTful==0.3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
Pillow==9.5.0
pytesseract==0.3.10
python-dotenv
pytz==2023.3
six
Werkzeug
zipp
